services/stt_service/pyproject.toml
[project]
name = "stt-service"
version = "0.1.0"
description = "Speech to Text GPU Service"
requires-python = ">=3.11"
dependencies = [
    "faster-whisper>=1.2.1",
    "grpcio-tools>=1.76.0",
    "rag-shared",
    "torch>=2.9.1",
]

[tool.uv.sources]
rag-shared = { workspace = true }
